v20.6.beta.2

+++ src/data/PolarODIM.h
@@ -249,8 +249,7 @@ public:
 	}
 
 
-	// defaultRange
-	static int defaultRange;
+	static double defaultRange;
 
 private:
